# [MXNET-1417][Performance] Caching Dynamic Shape Checking Result