@larroy 2019-06-18
@junrushao1994 would have been nice that you tagged me for review as well, since I was looking into this issue. The PR looks good to me anyways.