@larroy 2019-06-18
So the root cause of the degradation was the additional shape inferencing passes?