@larroy 2019-06-18
thanks for fixing this, and it's cool that your orginal commit enables dynamic shape.