@junrushao1994 2019-06-18
> So the root cause of the degradation was the additional shape inferencing passes?

Yep, the logic after these line is somewhat bulky.